+++ lib/json_matchers.rb
@@ -1,5 +1,4 @@
 require "json_matchers/version"
-require "json_matchers/configuration"
 require "json_matchers/matcher"
 require "json_matchers/errors"