+++ spec/json_matchers/configuration_spec.rb
@@ -1,9 +1,0 @@
-describe JsonMatchers, ".configure" do
-  it "ignores :record_errors configuration" do
-    with_options(record_errors: false) do
-      configured_options = JsonMatchers.configuration.options
-
-      expect(configured_options).to include(record_errors: true)
-    end
-  end
-end